requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
base==0.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dotenv==0.9.9
fastapi==0.116.1
field==0.2.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
iso8601==2.1.0
optional==0.0.1
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypika-tortoise==0.6.2
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
security==1.3.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
